feat: add expand and collapse button components to CoursePage test results bar
The expand button component and the collapse button component have been added to the CoursePage test results bar. These components are used to expand and collapse the test results bar respectively. They are implemented using the Glint component framework.

Additionally, a new data-test-bottom-section attribute has been added to the test results bar container element, and the clickOnBottomSection method has been added to the testResultsBar object in the coursePage module. This allows clicking on the bottom section of the test results bar to trigger the expand or collapse action.

In the acceptance test for the course page's view test results feature, the clickOnBottomSection method is now used instead of the click method to trigger the expand or collapse action on the test results bar.