Update AlertComponent and AutofixSectionComponent
The changes in the AlertComponent include adding dark mode styles for the success and error container colors. The changes in the AutofixSectionComponent include adding a new feature called Autofix, which automatically fixes issues in code. It includes a description of how it works and a "Start Autofix" button. If there is an error in creating the autofix request, an error message is displayed.

The RepositoryModel and SubmissionModel now have a relationship with the AutofixRequestModel.

Commit message: Update AlertComponent and AutofixSectionComponent