fix: update type declaration for unvote method in CourseExtensionIdeaModel
The unvote method in CourseExtensionIdeaModel was previously declared using the "unvote!" syntax, which is incorrect. It has been updated to use the correct syntax "declare unvote" to properly declare the method. Additionally, the "@ts-ignore" comment has been replaced with "@ts-expect-error" to indicate that an error is expected in this case.