feat: Update CopyableCode component to support dynamic background color based on prop
Removed hardcoded background color class and added conditional class based on value of props.backgroundColor in the CopyableCode component in copyable-code.hbs. Also added backgroundColor property to the Args interface in copyable-code.ts for type safety.
The changes were made to enhance the flexibility of the CopyableCode component by allowing the background color to be dynamically set based on the value passed in the props.